+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ### Added
 - Optional hiding offers dedicated for bundle (@goreck888)
 - Possibility to tag offers and bundles (@goreck888)
+- User suggestion input for bundle capabilities of goals values (@goreck888)
 
 ### Changed
 - Bundle linking available for all types of offers (@goreck888)